+++ app-wearos/src/main/java/host/stjin/anonaddy/utils/FavoriteAliasHelper.kt
@@ -5,17 +5,17 @@ import host.stjin.anonaddy.service.BackgroundWorkerHelper
 import host.stjin.anonaddy_shared.managers.SettingsManager
 
 class FavoriteAliasHelper(private val context: Context) {
-    private val settingsManager = SettingsManager(true, context)
+    private val encryptedSettingsManager = SettingsManager(true, context)
 
     fun getFavoriteAliases(): MutableSet<String>? {
-        return settingsManager.getStringSet(SettingsManager.PREFS.WEAROS_FAVORITE_ALIASES)
+        return encryptedSettingsManager.getStringSet(SettingsManager.PREFS.WEAROS_FAVORITE_ALIASES)
     }
 
 
     fun removeAliasAsFavorite(alias: String) {
         val aliasList = getFavoriteAliases()
         aliasList?.remove(alias)
-        aliasList?.let { settingsManager.putStringSet(SettingsManager.PREFS.WEAROS_FAVORITE_ALIASES, it) }
+        aliasList?.let { encryptedSettingsManager.putStringSet(SettingsManager.PREFS.WEAROS_FAVORITE_ALIASES, it) }
 
         // Since an alias was removed from the favorites, call scheduleBackgroundWorker. This method will schedule the service if its still required
         BackgroundWorkerHelper(context).scheduleBackgroundWorker()
@@ -26,7 +26,7 @@ class FavoriteAliasHelper(private val context: Context) {
         if (aliasList != null) {
             if (aliasList.size < 3) {
                 aliasList.add(alias)
-                aliasList.let { settingsManager.putStringSet(SettingsManager.PREFS.WEAROS_FAVORITE_ALIASES, it) }
+                aliasList.let { encryptedSettingsManager.putStringSet(SettingsManager.PREFS.WEAROS_FAVORITE_ALIASES, it) }
 
                 // Since an alias was added to the favorites, call scheduleBackgroundWorker. This method will schedule the service if its required
                 BackgroundWorkerHelper(context).scheduleBackgroundWorker()
@@ -38,6 +38,6 @@ class FavoriteAliasHelper(private val context: Context) {
     }
 
     fun clearFavoriteAliases() {
-        settingsManager.removeSetting(SettingsManager.PREFS.WEAROS_FAVORITE_ALIASES)
+        encryptedSettingsManager.removeSetting(SettingsManager.PREFS.WEAROS_FAVORITE_ALIASES)
     }
 }